\name{hbm-package}
\alias{hbm-package}

\docType{package}
\title{
Hierarchical Block Matrix Analysis 
}
\description{
A package for building hierarchical block matrices from association matrices and for performing multi-scale analysis. It specifically targets chromatin contact maps, generated from high-throughput chromosome conformation capture data, such as 5C and Hi-C, and provides methods for detecting movements and for computing chain hierarchy and region communicability across scales.
}
\details{
\tabular{ll}{
Package: \tab hbm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-25\cr
License: \tab GPL (>=2)\cr
}
Get started with \code{hbm}'s tutorials at: \href{http://www.cl.cam.ac.uk/~ys388/hbm/}{http://www.cl.cam.ac.uk/~ys388/hbm/}
}
\author{
Yoli Shavit\cr
Maintainer: <ys388@cam.ac.uk>

}
\references{
\code{hbm}'s website: \href{http://www.cl.cam.ac.uk/~ys388/hbm/}{http://www.cl.cam.ac.uk/~ys388/hbm/}
}


